refactor: split the backtest service into focused modules
BacktestService was one 1400-line class holding the definition registry,
Paris calendar math, daily regime analytics, trade statistics and the
simulation engine. Extract the parts that carry no orchestration into
their own modules under a new api/services/backtest package:

- definitions.py  the four registered backtests + parameter merging
- calendar.py     9h reference window and session end as naive UTC
- candles.py      the shared candle_date guard
- analytics.py    MA50 slope, ADX and overnight gap, with the lookahead
                  guard and newest-first ordering factored into one helper
- statistics.py   trade aggregation into a BacktestSummary

These were @staticmethods calling each other through an explicit
BacktestService. prefix - free functions in all but name. Behavior is
unchanged; the golden snapshot is untouched.

Tests move with the code, one file per module, and share fixtures through
tests/api/services/backtest/helpers.py.
